% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes.R
\name{dupes_find}
\alias{dupes_find}
\title{Identify duplicate entries}
\usage{
dupes_find(x, match_cols, approx_match = FALSE, string_dist = 5,
  min_length = 20, simplify_match = TRUE)
}
\arguments{
\item{x}{The dataset in which duplicate entries will be identified}

\item{match_cols}{Column(s) that will be used to search for duplicate
records}

\item{approx_match}{Whether to perform a duplicate search using
string distances or exact values}

\item{string_dist}{When using approximate matching, the string
distance cutoff at which records will be assumed duplicated}

\item{min_length}{The minimum length for the combined matching string
produced by \code{match_cols} at which a record will be considered for
matching}

\item{simplify_match}{Whether to perform duplicate searches after removing
all non alpha-numeric characters from the reference string generated from
\code{match_cols}}
}
\value{
An updated version of \code{x}, with one column specifying the
final string used to search for duplicates (\code{matching_col})
and another column containing unique identifiers for each set of
duplicates (\code{match_ID}).
}
\description{
For a tabular set of publication records, identifies potential sets of
duplicate entries and labels them with a unique identifier.
}
\examples{
\dontrun{
test <- rbind(form_mm_recs, form_mm_recs)
test <- dupes_find(test, c(1, 3))
dupes <- dupes_return(test)
out <- dupes_rm(test)
}
}
