% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes_rm_1way.R
\name{dupes_rm_1way}
\alias{dupes_rm_1way}
\title{Remove previously documented publication records}
\usage{
dupes_rm_1way(x, ignore_IDs = NULL)
}
\arguments{
\item{x}{The output of \code{dupes_find_1way}}

\item{ignore_IDs}{A set of match IDs that will be ignored when removing
duplicate records. These should be identified by manually inspecting
the output of \code{dupes_return}.}
}
\value{
An updated version of \code{x}, with duplicate records removed.
}
\description{
Takes output of \code{dupes_find_1way}, and returns a set of combined records
with duplicate entries removed.
}
\details{
After running \code{dupes_find_1way}, records from \code{old} and \code{new}
are combined and pairs of entries matched between \code{old} and \code{new}
are marked with identical match IDs specified in the \code{match_ID} column.
This function removes the duplicate entry that originated from the \code{new}
set of records.
}
\examples{
\dontrun{
old <- form_mm_recs[1:10, ]
new <- form_mm_recs[8:12, ]
test <- dupes_find_1way(old, new, c(1, 3))
dupes <- dupes_return(test)
out <- dupes_rm_1way(test)
}
}
