% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dupes_1way.R
\name{find_dupes_1way}
\alias{find_dupes_1way}
\title{Identify previously documented publication records}
\usage{
find_dupes_1way(old, new, match_cols, approx_match = FALSE,
  string_dist = 10, min_length = 20, simplify_match = TRUE)
}
\arguments{
\item{old}{The previous set of publication records}

\item{new}{The new set of publication records}

\item{match_cols}{Column(s) that will be used to search for duplicate
records}

\item{approx_match}{Whether to perform a duplicate search using
string distances or exact values}

\item{string_dist}{When using approximate matching, the string
distance cutoff at which records will be assumed duplicated}

\item{min_length}{The minimum string length for \code{match_cols} at which
a record will be considered when searching for duplicates}

\item{simplify_match}{Whether to perform duplicate searches after removing
all non alpha-numeric characters from the reference string generated from
\code{match_cols}}
}
\value{
The combined records from \code{old} and \code{new} records from new
found in old, indicated by paired match_IDs.
duplicates (\code{match_ID}).
}
\description{
For two sets of publication records, one old and one new, combines old and
new records and marks instances where new records are duplicated in the old
set of records.
}
\examples{
\dontrun{
old <- form_mm_recs[1:10, ]
new <- form_mm_recs[8:12, ]
test <- find_dupes_1way(old, new, c(1, 3))
dupes <- return_duplicates(test)
out <- remove_dupes_1way(test)
}
}
