% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Identify duplicate entries}
\usage{
find_duplicates(x, match_cols, approx_match = FALSE, string_dist = 10,
  min_length = 20, simplify_match = TRUE)
}
\arguments{
\item{x}{A dataset in which duplicate entries will be identified}

\item{match_cols}{Column(s) that will be used to generate a search string for
duplicate identification. If more than one column is
specified, then the search string will be formed from the
concatenation of values from column specified in \code{match_cols}.}

\item{approx_match}{If set to \code{TRUE}, performs a duplicate search using
string distances rather than exact values.}

\item{string_dist}{When performing approximate matching, is is the distance
cutoff at which entries will be considered matched.}

\item{min_length}{The minimum length for values in \code{match_cols}
such that rows with \code{match_cols} shorter than the values specified by
this argument will be protected from identification as a duplicate.
This helps ensure that aberrant duplicates are not identified due to
very short or empty values within \code{match_cols}.}

\item{simplify_match}{If \code{TRUE}, modifies the matching string by
removing all alpha-numeric characters and setting the remaining string to
lower-case.}
}
\value{
An updated version of \code{x}, with extra columns specifying the
final string used to search for duplicates (contained within a new column
called \code{matching_col}) and a unique identifier for each set of
duplicates (contained within a new column called \code{match_ID})
that that will share the same value for any identified duplicates.
}
\description{
Identifies duplicate entries in a set of tabular data using one or more
columns, then labels matches by a unique match identifier.
}
\examples{
\dontrun{
}
}
