% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{form_mm_recs}
\alias{form_mm_recs}
\title{Myeloma publications records formatted into a \code{data.frame}}
\format{An object of class \code{data.frame} with 98 rows and 14 columns.}
\usage{
form_mm_recs
}
\description{
Myeloma publications records formatted into a \code{data.frame}
}
\keyword{datasets}
