% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_scopus.R
\name{format_scopus}
\alias{format_scopus}
\title{Builds a \code{data.frame} containing publication info from an Scopus
query}
\usage{
format_scopus(x)
}
\arguments{
\item{x}{A \code{data.frame} constructed from manually downloaded Scopus
data.}
}
\value{
A \code{data.frame} formatted for combination with other sources of
bibliographic data.
}
\description{
This function takes the manually downloaded publication data from a Scopus
query and formats it so that it can be combined with bibliographic data
from other sources.
}
\examples{
\dontrun{
scopus_data <- read.csv("your_scopus_data.csv", stringsAsFactors = FALSE)
form_s_data <- format_scopus(scopus_data)
}
}
\seealso{
\code{\link{format_RISmed}}
}
