% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy.R
\name{fuzzy_in}
\alias{fuzzy_in}
\title{Find approximately matching values}
\usage{
fuzzy_in(x, y, string_dist = 5, protect_length = 5)
}
\arguments{
\item{x}{The values to be matched}

\item{y}{The values to be matched against}

\item{string_dist}{A minimum string distance}

\item{protect_length}{A minimum length a which strings will be considered}
}
\value{
A logical vector indicate values in \code{x} duplicated in \code{y}
}
\description{
Finds values from \code{x} that are approximately duplicated \code{y}
}
\examples{
\dontrun{
fuzzy_in(c("Apples", "Oranges", "Bananas"),
         c("Apple", "Oranges", "Bananas")
         string_dist = 3, protect_length = 3)
}
}
