% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myeloma_recs}
\alias{myeloma_recs}
\title{A \code{Medline} object containing myeloma publication records}
\format{An object of class \code{Medline} of length 1.}
\usage{
myeloma_recs
}
\description{
A \code{Medline} object containing myeloma publication records
}
\keyword{datasets}
