% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes_rm.R
\name{remove_dupes}
\alias{remove_dupes}
\title{Remove duplicate entries}
\usage{
remove_dupes(x, db_pref = NULL, ignore_IDs = NULL)
}
\arguments{
\item{x}{The output of \code{dupes_find}}

\item{db_pref}{The preferred database for a reference.}

\item{ignore_IDs}{A set of match IDs that will be ignored when removing
duplicate records. These should be identified by manually inspecting
the output of \code{dupes_return}.}
}
\value{
An updated version of \code{x}, with duplicate records removed.
}
\description{
Takes output of \code{dupes_find}, and returns a set of records with
duplicate entries removed.
}
\details{
After running \code{dupes_find}, duplicate sets of records will contain
identical match IDs specified in the \code{match_ID} column. This
function will remove all but one entry from each set of duplicates. The
preserved entry is selected based on the user's preferred database
(\code{db_pref}) and the entry within the duplicate set with the
least number of blank cells.
}
\examples{
\dontrun{
test <- rbind(form_mm_recs, form_mm_recs)
test <- dupes_find(test, c(1, 3))
dupes <- dupes_return(test)
out <- remove_dupes(test)
}
}
