% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dupes_1way.R
\name{remove_dupes_1way}
\alias{remove_dupes_1way}
\title{Remove previously documented publication records}
\usage{
remove_dupes_1way(x, ignore_IDs = NULL)
}
\arguments{
\item{x}{The output of \code{find_dupes_1way}}

\item{ignore_IDs}{A set of match IDs that will be ignored when removing
duplicate records. These should be identified by manually inspecting
the output of \code{return_duplicates_df}.}
}
\value{
An updated version of \code{x}, with duplicate records removed.
}
\description{
Takes output of \code{find_dupes_1way}, and returns a set of combined records
with duplicate entries removed.
}
\details{
After running \code{find_dupes_1way}, records from \code{old} and \code{new}
are combined and pairs of entries matched between \code{old} and \code{new}
are marked with identical match IDs specified in the \code{match_ID} column.
This function removes the duplicate entry that originated from the \code{new}
set of records.
}
\examples{
\dontrun{
}
}
