% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{return_duplicates_df}
\alias{return_duplicates_df}
\title{Returns duplicated entries}
\usage{
return_duplicates_df(x)
}
\arguments{
\item{x}{The ouput of \code{find_duplicates}}
}
\value{
A \code{data.frame} containing only entries of the original dataset
identified as duplicate entries. This is meant to facilitate manual
inspection of potential duplicates. Aberrantly identified duplicates
may subsequently be spared when running \code{remove_duplicates}
by indicating their match IDs in \code{remove_duplicates}'s
\code{ignore_IDs} argument.
}
\description{
Takes output generated by \code{find_duplicates}, and returns a
\code{data.frame} containing only those entries of the original dataset
identified as duplicate entries.
}
\examples{
\dontrun{
}
}
