% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_fulltxt_list}
\alias{screen_fulltxt_list}
\title{Create a list to screen full texts}
\usage{
screen_fulltxt_list(x, reviewers, cols = "TITLE")
}
\arguments{
\item{x}{A set of publication records without duplicates}

\item{reviewers}{Names of people that will screen full texts}

\item{cols}{Columns from x to include in the full-text screeners}
}
\value{
A list of of \code{data.frame}s, formatted for each reviewer to
screen abstracts
}
\description{
Creates a list of \code{data.frame}s to screen full texts
}
\examples{
\dontrun{
screening_list <- screen_abst_list(form_mm_recs, c("you", "me"))
}
}
