% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_list_abst}
\alias{screen_list_abst}
\title{Create a list of \code{data.frame}s to screen abstracts}
\usage{
screen_list_abst(x, reviewers, title_col = "TITLE",
  abstract_col = "ABSTRACT")
}
\arguments{
\item{x}{A set of publication records without duplicates}

\item{reviewers}{Names of people that will screen abstracts}

\item{title_col}{Column name containing the publication title}

\item{abstract_col}{Column name containing the publication abstract}
}
\value{
A list of of \code{data.frame}s, formatted for each reviewer to
screen abstracts
}
\description{
Creates a list of \code{data.frame}s to screen abstracts
}
\examples{
\dontrun{
screening_list <- screen_list_abst(form_mm_recs, c("you", "me"))
}
}
