% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscreen.R
\name{screen_read}
\alias{screen_read}
\title{Reads completed abstract screening .csv}
\usage{
screen_read(dir = "../intermediate_data/", ref_table)
}
\arguments{
\item{dir}{The directory containing complete screening forms}

\item{ref_table}{The original set of records from which screening sheets
were collected. Note: this must be unaltered from the point at which
screening forms were first generated with \code{screen_make_list}.}
}
\value{
A \code{data.frame} with the combined results of all screening
sheets, plus the publication data from \code{ref_table}
}
\description{
Takes a series of completed abstract screening forms and assembles them
into a \code{data.frame}
}
\examples{
\dontrun{
screening_list <- screen_make_list(form_mm_recs, c("you", "me"))
screen_write(screening_list, dir = "./")
# Abstracts should be screened at this point
combined_screening_forms <- screen_read("./", form_mm_recs)
}
}
