% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_read_fulltxt}
\alias{screen_read_fulltxt}
\title{Reads completed full-text screening .csv files}
\usage{
screen_read_fulltxt(ref_table, dir = "../intermediate_data/")
}
\arguments{
\item{ref_table}{The original set of records from which screening sheets
were collected. Note: this must be unaltered from the point at which
screening forms were first generated with \code{screen_list_abst}.}

\item{dir}{The directory containing complete screening forms}
}
\value{
A \code{data.frame} with the combined results of all screening
sheets, plus the publication data from \code{ref_table}
}
\description{
Takes a series of completed full-text screening forms and assembles them
into a \code{data.frame}
}
\examples{
\dontrun{
screening_list <- screen_list_fulltxt(form_mm_recs, c("you", "me"))
screen_list_fulltxt(screening_list, dir = "./")
# Full-text articles should be screened at this point
combined_screening_forms <- screen_read_fulltxt("./", form_mm_recs)
}
}
