% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscreen.R
\name{screen_write}
\alias{screen_write}
\title{Create .csv files for abstract screening}
\usage{
screen_write(x, dir = "../intermediate_data/")
}
\arguments{
\item{x}{The list created by \code{screen_make_list}}

\item{dir}{The desired directory into which .csv files will be placed}
}
\description{
Takes a the output of \code{screen_make_list} and creates a series of
.csv files that may be distributed to the individuals that will screen
abstracts
}
\examples{
\dontrun{
screening_list <- screen_make_list(form_mm_recs, c("you", "me"))
screen_write(screening_list, dir = "./")
}
}
