% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_write_abst}
\alias{screen_write_abst}
\title{Create .csv files for abstract screening}
\usage{
screen_write_abst(x, dir = "../intermediate_data/")
}
\arguments{
\item{x}{The list created by \code{screen_list_abst}}

\item{dir}{The desired directory into which .csv files will be placed}
}
\description{
Takes a the output of \code{screen_list_abst} and creates a series of
.csv files that may be distributed to the individuals that will screen
abstracts
}
\examples{
\dontrun{
screening_list <- screen_list_abst(form_mm_recs, c("you", "me"))
screen_write_abst(screening_list, dir = "./")
}
}
