% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_data}
\alias{update_data}
\title{Update a revised publication table with previously extracted information}
\usage{
update_data(empty, populated, match_cols, approx_match = FALSE,
  string_dist = 10, min_length = 20, simplify_match = TRUE)
}
\arguments{
\item{empty}{The new set of publication records}

\item{populated}{Previously recorded publication data}

\item{match_cols}{Column(s) that will be used to match
\code{empty} and \code{populated}}

\item{approx_match}{Whether to use string distances or exact values when
matching records.}

\item{string_dist}{When using approximate matching, the string
distance cutoff at which records will be matched.}

\item{min_length}{The minimum string length for \code{match_cols} at which
a record will be considered when matching records.}

\item{simplify_match}{Whether to perform matching on strings composed from
\code{match_cols}, but with non alpha-numeric values removed.}
}
\value{
An updated version of \code{empty}, which will be updated where
matches to \code{populated} (based on \code{match_cols}) are made.
}
\description{
Takes a new set of publication records populated and updates with
previously recorded (often manually acquired) data.
}
\examples{
\dontrun{
empty <- data.frame(a = c("Apples", "Oranges", "Bananas"),
                    b = c("Granny", "Florida", "Chiquita"),
                    c = c("", "", ""),
                    d = c("", "", ""),
                    stringsAsFactors = FALSE)

update <- data.frame(a = c("Apples", "Oranges"),
                     b = c("Granny", "Florida"),
                     c = c("Red", ""),
                     d = c("Green", ""),
                     stringsAsFactors = FALSE)

update_data(empty, update, c("a", "b"), min_length = 5)
}
}
