% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract_screening.R
\name{write_abstract_screeners}
\alias{write_abstract_screeners}
\title{Create .csv files for abstract screening}
\usage{
write_abstract_screeners(x, dir = "../intermediate_data/")
}
\arguments{
\item{x}{The list created by \code{make_screening_list}}

\item{dir}{The desired directory into which .csv files will be placed}
}
\description{
Takes a the output of \code{make_screening_list} and creates a series of
.csv files that may be distributed to the individuals that will screen
abstracts
}
\examples{
\dontrun{
screening_list <- make_screening_list(form_mm_recs, c("you", "me"))
write_abstract_screeners(screening_list, dir = "./")
}
}
