% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate entries}
\usage{
remove_duplicates(x, db_pref = NULL, ignore_IDs = NULL)
}
\arguments{
\item{x}{The output of \code{find_duplicates}}

\item{db_pref}{The preferred database for a reference.}

\item{ignore_IDs}{A set of match IDs that will be ignored when removing
duplicate records. These should be identified by manually inspecting
the output of \code{return_duplicates_df}.}
}
\value{
An updated version of \code{x}, with duplicate records removed.
}
\description{
Takes output of \code{find_duplicates}, and returns a set of records with
duplicate entries removed.
}
\details{
After running \code{find_duplicates}, duplicate sets of records will contain
identical match IDs specified in the \code{match_ID} column. This
function will remove all but one entry from each set of duplicates. The
preserved entry is selected based on the user's preferred database
(\code{db_pref}) and the entry within the duplicate set with the
least number of blank cells.
}
\examples{
\dontrun{
}
}
